+++ tests/fixest_tests.R
@@ -1141,6 +1141,16 @@ est_feols = feols(y ~ x | grp + tm, data=d)
 test(se(est_feols, se = "st")["x"], se(est_lm)["x"])
 
 #
+# Heteroskedasticity-robust
+#
+
+se_white_lm_HC1 = sqrt(vcovHC(est_lm, type = "HC1")["x", "x"])
+se_white_lm_HC0 = sqrt(vcovHC(est_lm, type = "HC0")["x", "x"])
+
+test(se(est_feols, se = "hetero"), se_white_lm_HC1)
+test(se(est_feols, se = "hetero", ssc = ssc(adj = FALSE, cluster.adj = FALSE)), se_white_lm_HC0)
+
+#
 # Clustered
 #
 
@@ -1153,16 +1163,6 @@ test(se(est_feols, ssc = ssc(fixef.K = "full")), se_CL_grp_lm_HC1)
 test(se(est_feols, ssc = ssc(adj = FALSE, fixef.K = "full")), se_CL_grp_lm_HC0)
 
 #
-# Heteroskedasticity-robust
-#
-
-se_white_lm_HC1 = sqrt(vcovHC(est_lm, type = "HC1")["x", "x"])
-se_white_lm_HC0 = sqrt(vcovHC(est_lm, type = "HC0")["x", "x"])
-
-test(se(est_feols, se = "hetero"), se_white_lm_HC1)
-test(se(est_feols, se = "hetero", ssc = ssc(adj = FALSE, cluster.adj = FALSE)), se_white_lm_HC0)
-
-#
 # Two way
 #
 
@@ -1172,13 +1172,82 @@ se_CL_2w_feols = se(est_feols, se = "twoway")
 
 test(se(est_feols, se = "twoway", ssc = ssc(fixef.K = "full", cluster.df = "conv")), se_CL_2w_lm)
 
+# 
+# HC2/HC3
+#
+
+# HC2/HC3
+base = iris
+base$w = runif(nrow(base))
+
+est_feols = feols(Sepal.Length ~ Sepal.Width | Species, base)
+est_lm = lm(Sepal.Length ~ Sepal.Width + factor(Species), base)
+
+test(
+  vcov(est_feols, "hc2", ssc = ssc(adj = FALSE, cluster.adj = FALSE)),
+  sandwich::vcovHC(est_lm, "HC2")["Sepal.Width", "Sepal.Width"]
+)
+
+test(
+  vcov(est_feols, "hc3", ssc = ssc(adj = FALSE, cluster.adj = FALSE)),
+  sandwich::vcovHC(est_lm, "HC3")["Sepal.Width", "Sepal.Width"]
+)
+
+# HC2/HC3 with weights
+base = iris
+base$w = runif(nrow(base))
+
+est_feols = feols(Sepal.Length ~ Sepal.Width | Species, base, weights = ~ w)
+est_lm = lm(Sepal.Length ~ Sepal.Width + factor(Species), base, weights = base$w)
+
+test(
+  vcov(est_feols, "hc2", ssc = ssc(adj = FALSE, cluster.adj = FALSE)),
+  sandwich::vcovHC(est_lm, "HC2")["Sepal.Width", "Sepal.Width"]
+)
+
+test(
+  vcov(est_feols, "hc3", ssc = ssc(adj = FALSE, cluster.adj = FALSE)),
+  sandwich::vcovHC(est_lm, "HC3")["Sepal.Width", "Sepal.Width"]
+)
+
+# HC2/HC3 with GLM
+base$Sepal.Length = floor(base$Sepal.Length)
+est_feglm = feglm(
+  Sepal.Length ~ Sepal.Width | Species, base, 
+  "poisson", weights = ~ w
+)
+est_glm = glm(
+  Sepal.Length ~ Sepal.Width + factor(Species), base, 
+  family = poisson(), weights = base$w
+)
+
+test(
+  vcov(est_feglm, "hc2", ssc = ssc(adj = FALSE, cluster.adj = FALSE)),
+  sandwich::vcovHC(est_glm, "HC2")["Sepal.Width", "Sepal.Width"]
+)
+test(
+  vcov(est_feglm, "hc3", ssc = ssc(adj = FALSE, cluster.adj = FALSE)),
+  sandwich::vcovHC(est_glm, "HC3")["Sepal.Width", "Sepal.Width"]
+)
+
+# Fail when P_ii = 1
+base$Species = as.character(base$Species)
+base[1, "Species"] = "foo"
+
+est_pii_singular = feols(Sepal.Length ~ Sepal.Width | Species, base)
+
+test(
+  vcov(est_pii_singular, "hc2"), "err"
+)
+
+
 #
 # Checking the calls work properly
 #
 
 data(trade)
 
-est_pois = femlm(Euros ~ log(dist_km)|Origin+Destination, trade)
+est_pois = femlm(Euros ~ log(dist_km) | Origin + Destination, trade)
 
 se_clust = se(est_pois, se = "cluster", cluster = "Product")
 test(se(est_pois, cluster = trade$Product), se_clust)